Feat/Fix: Add public metadata and decorator. Move directory
