+++ src/authentication/user/v1/user.service.ts
@@ -4,4 +4,8 @@ import { UserAuthV1Repository } from './user.repository';
 @Injectable()
 export class UserAuthV1Service {
   constructor(private readonly repository: UserAuthV1Repository) {}
+
+  async signup() {}
+
+  async signin() {}
 }
