+++ src/decorator/index.ts
@@ -1,1 +1,3 @@
-export * from './pagination.decorator';
+export * from './pagination/pagination.decorator';
+export * from './public/public.decorator';
+export * from './swagger';
