+++ src/decorator/swagger/index.ts
@@ -1,0 +1,2 @@
+export * from './api-multiple-body.decorator';
+export * from './api-multiple-response.decorator';
